# internal/rpc/openapi.yaml
openapi: "3.0.3"
info:
  title: DRANA Node RPC
  description: |
    JSON HTTP API for the DRANA blockchain node. Use this to query chain state,
    submit transactions, and inspect blocks.

    All amounts are in **microdrana** (1 DRANA = 1,000,000 microdrana).
    Hex values are lowercase, no `0x` prefix.
    Addresses use the `drana1` display format.
  version: "1.0.0"
...
paths:
...
  /v1/blocks/latest:
    get:
      tags: [Chain]
      summary: Get latest block
      parameters:
        - name: full
          in: query
          schema: { type: boolean, default: false }
          description: Include full transaction list
      responses:
        "200":
          description: Latest block
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Block"
...
  /v1/posts:
    get:
      tags: [Posts]
      summary: List posts
      description: Paginated list of top-level posts. Replies excluded by default.
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, default: 20, maximum: 100 }
        - name: author
          in: query
          schema: { type: string }
          description: Filter by author address
        - name: channel
          in: query
          schema: { type: string }
          description: Filter by channel name
        - name: includeReplies
          in: query
          schema: { type: boolean, default: false }
          description: Include replies in results
      responses:
        "200":
          description: Post list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostList"
...
  /v1/transactions:
    post:
      tags: [Transactions]
      summary: Submit transaction
      description: Submit a signed transaction to the network.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitTxRequest"
      responses:
        "200":
          description: Submission result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmitTxResponse"
...
components:
  schemas:
...
    Block:
      type: object
      properties:
        height: { type: string, description: "uint64 encoded as string" }
        hash: { type: string, description: "64-char hex" }
        prevHash: { type: string }
        proposerAddr: { type: string, description: "drana1..." }
        timestamp: { type: integer, format: int64 }
        stateRoot: { type: string }
        txRoot: { type: string }
        txCount: { type: integer }
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
...
    Post:
      type: object
      properties:
        postId: { type: string, description: "64-char hex" }
        author: { type: string }
        text: { type: string }
        channel: { type: string }
        parentPostId: { type: string, description: "Empty for top-level posts" }
        createdAtHeight: { type: string, description: "uint64 encoded as string" }
        createdAtTime: { type: integer, format: int64 }
        totalStaked: { type: string, description: "uint64 encoded as string, active staked microdrana" }
        totalBurned: { type: string, description: "uint64 encoded as string, cumulative fees burned" }
        stakerCount: { type: string, description: "uint64 encoded as string" }
        withdrawn: { type: boolean }
        boostCount: { type: string, description: "uint64 encoded as string" }
    PostList:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        totalCount: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
    SubmitTxRequest:
      type: object
      required: [type, sender, amount, nonce, signature, pubKey]
      properties:
        type:
          type: string
          enum: [transfer, create_post, boost_post, unstake_post, register_name, stake, unstake]
        sender: { type: string, description: "drana1... address" }
        recipient: { type: string, description: "For transfers" }
        postId: { type: string, description: "For boost_post (64-char hex)" }
        parentPostId: { type: string, description: "For replies (64-char hex)" }
        text: { type: string, description: "For create_post or register_name" }
        channel: { type: string, description: "For create_post" }
        amount: { type: string, description: "uint64 encoded as string, microdrana" }
        nonce: { type: string, description: "uint64 encoded as string" }
        signature: { type: string, description: "Ed25519 signature (hex)" }
        pubKey: { type: string, description: "Ed25519 public key (hex)" }
    SubmitTxResponse:
      type: object
      properties:
        accepted: { type: boolean }
        txHash: { type: string }
        error: { type: string }
    Transaction:
      type: object
      properties:
        hash: { type: string }
        type: { type: string }
        sender: { type: string }
        recipient: { type: string }
        postId: { type: string }
        parentPostId: { type: string }
        text: { type: string }
        channel: { type: string }
        amount: { type: string, description: "uint64 encoded as string" }
        nonce: { type: string, description: "uint64 encoded as string" }
        blockHeight: { type: string, description: "uint64 encoded as string" }
...